+++ vct.h
@@ -37,7 +37,7 @@
 #endif
 
 typedef struct {
-  int length;
+  off_t length;
   Float *data;
   bool dont_free;
 } vct;
@@ -46,18 +46,18 @@ void mus_vct_init(void);
 bool mus_vct_p(XEN obj);
 int mus_vct_print_length(void);
 void mus_vct_set_print_length(int len);
-XEN mus_array_to_list(Float *arr, int i, int len);
+XEN mus_array_to_list(Float *arr, off_t i, off_t len);
 char *mus_vct_to_string(vct *v);
 bool mus_vct_equalp(vct *v1, vct *v2);
 char *mus_vct_to_readable_string(vct *v);
-vct *mus_vct_make(int len);
+vct *mus_vct_make(off_t len);
 vct *mus_vct_free(vct *v);
 vct *mus_vct_copy(vct *vc);
 
 XEN xen_list_to_vct(XEN lst);
 vct *xen_to_vct(XEN arg);
-XEN xen_make_vct(int len, Float *data);
-XEN xen_make_vct_wrapper(int len, Float *data);
+XEN xen_make_vct(off_t len, Float *data);
+XEN xen_make_vct_wrapper(off_t len, Float *data);
 
 #define XEN_TO_VCT(arg) ((vct *)XEN_OBJECT_REF(arg))
 #define MUS_VCT_P(arg) mus_vct_p(arg)
